# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Sistem Koperasi Merah Putih API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /koperasi:
    get:
      operationId: listKoperasi
      tags: [koperasi]
      summary: List koperasi
      parameters:
        - in: query
          name: status
          schema:
            type: string
        - in: query
          name: provinsi
          schema:
            type: string
      responses:
        "200":
          description: List koperasi
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Koperasi"
    post:
      operationId: createKoperasi
      tags: [koperasi]
      summary: Daftarkan koperasi baru
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/KoperasiInput"
      responses:
        "201":
          description: Koperasi didaftarkan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Koperasi"
  /koperasi/{id}:
    get:
      operationId: getKoperasi
      tags: [koperasi]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Detail koperasi
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Koperasi"
    patch:
      operationId: updateKoperasi
      tags: [koperasi]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/KoperasiUpdate"
      responses:
        "200":
          description: Koperasi diupdate
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Koperasi"
  /koperasi/{id}/verifikasi:
    post:
      operationId: verifikasiKoperasi
      tags: [koperasi]
      summary: Verifikasi koperasi (Super Admin)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifikasiInput"
      responses:
        "200":
          description: Status verifikasi diperbarui
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Koperasi"

  # ─── ANGGOTA ─────────────────────────────────────────────────────────────────
...
  /pinjaman:
    get:
      operationId: listPinjaman
      tags: [pinjaman]
      summary: List pengajuan pinjaman
      parameters:
        - in: query
          name: anggotaId
          schema:
            type: integer
        - in: query
          name: koperasiId
          schema:
            type: integer
        - in: query
          name: status
          schema:
            type: string
      responses:
        "200":
          description: List pinjaman
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Pinjaman"
    post:
      operationId: createPinjaman
      tags: [pinjaman]
      summary: Ajukan pinjaman
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PinjamanInput"
      responses:
        "201":
          description: Pinjaman diajukan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Pinjaman"
  /pinjaman/{id}:
    get:
      operationId: getPinjaman
      tags: [pinjaman]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Detail pinjaman
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Pinjaman"
  /pinjaman/{id}/setujui:
    post:
      operationId: setujuiPinjaman
      tags: [pinjaman]
      summary: Setujui atau tolak pinjaman
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetujuiPinjamanInput"
      responses:
        "200":
          description: Status pinjaman diperbarui
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Pinjaman"

  # ─── ANGSURAN ─────────────────────────────────────────────────────────────────
...
  /produk:
    get:
      operationId: listProduk
      tags: [produk]
      summary: List produk/stok
      parameters:
        - in: query
          name: unitUsahaId
          schema:
            type: integer
        - in: query
          name: kategori
          schema:
            type: string
      responses:
        "200":
          description: List produk
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Produk"
    post:
      operationId: createProduk
      tags: [produk]
      summary: Tambah produk
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProdukInput"
      responses:
        "201":
          description: Produk ditambah
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Produk"
  /produk/{id}:
    patch:
      operationId: updateProduk
      tags: [produk]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProdukUpdate"
      responses:
        "200":
          description: Produk diupdate
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Produk"

  # ─── TRANSAKSI ─────────────────────────────────────────────────────────────────
...
  /laporan/keuangan:
    get:
      operationId: getLaporanKeuangan
      tags: [laporan]
      summary: Laporan keuangan koperasi
      parameters:
        - in: query
          name: koperasiId
          schema:
            type: integer
        - in: query
          name: tahun
          schema:
            type: integer
        - in: query
          name: bulan
          schema:
            type: integer
      responses:
        "200":
          description: Laporan keuangan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LaporanKeuangan"
  /laporan/shu:
    get:
      operationId: getLaporanShu
      tags: [laporan]
      summary: Laporan SHU
      parameters:
        - in: query
          name: koperasiId
          schema:
            type: integer
        - in: query
          name: tahun
          schema:
            type: integer
      responses:
        "200":
          description: Laporan SHU
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LaporanShu"
  /laporan/audit:
    get:
      operationId: getLaporanAudit
      tags: [laporan]
      summary: Laporan audit pengawas
      parameters:
        - in: query
          name: koperasiId
          schema:
            type: integer
        - in: query
          name: tahun
          schema:
            type: integer
      responses:
        "200":
          description: Laporan audit
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LaporanAudit"

  # ─── DASHBOARD ────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Koperasi:
      type: object
      required: [id, nama, desa, kecamatan, kabupaten, provinsi, status, createdAt]
      properties:
        id:
          type: integer
        nama:
          type: string
        noBadanHukum:
          type: ["string", "null"]
        desa:
          type: string
        kecamatan:
          type: string
        kabupaten:
          type: string
        provinsi:
          type: string
        alamat:
          type: ["string", "null"]
        telepon:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, aktif, nonaktif, ditolak]
        tanggalBerdiri:
          type: ["string", "null"]
        jumlahAnggota:
          type: integer
        totalAset:
          type: number
        createdAt:
          type: string
    KoperasiInput:
      type: object
      required: [nama, desa, kecamatan, kabupaten, provinsi]
      properties:
        nama:
          type: string
        noBadanHukum:
          type: string
        desa:
          type: string
        kecamatan:
          type: string
        kabupaten:
          type: string
        provinsi:
          type: string
        alamat:
          type: string
        telepon:
          type: string
        email:
          type: string
        tanggalBerdiri:
          type: string
    KoperasiUpdate:
      type: object
      properties:
        nama:
          type: string
        noBadanHukum:
          type: string
        alamat:
          type: string
        telepon:
          type: string
        email:
          type: string
        status:
          type: string
    VerifikasiInput:
      type: object
      required: [status]
      properties:
        status:
          type: string
        catatan:
          type: string

    # ─── ANGGOTA ──────────────────────────────────────────────────────────────
...
    Pinjaman:
      type: object
      required: [id, anggotaId, jumlahPinjaman, bungaPersen, tenorBulan, status, tanggalPengajuan, createdAt]
      properties:
        id:
          type: integer
        anggotaId:
          type: integer
        namaAnggota:
          type: string
        jumlahPinjaman:
          type: number
        bungaPersen:
          type: number
        tenorBulan:
          type: integer
        angsuranPerBulan:
          type: number
        tujuan:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, disetujui, ditolak, lunas, macet]
        tanggalPengajuan:
          type: string
        tanggalDisetujui:
          type: ["string", "null"]
        tanggalJatuhTempo:
          type: ["string", "null"]
        catatanPengurus:
          type: ["string", "null"]
        sisaPinjaman:
          type: number
        createdAt:
          type: string
    PinjamanInput:
      type: object
      required: [anggotaId, jumlahPinjaman, tenorBulan, tujuan]
      properties:
        anggotaId:
          type: integer
        jumlahPinjaman:
          type: number
        tenorBulan:
          type: integer
        tujuan:
          type: string
    SetujuiPinjamanInput:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [disetujui, ditolak]
        bungaPersen:
          type: number
        catatanPengurus:
          type: string

    # ─── ANGSURAN ─────────────────────────────────────────────────────────────
...
    Produk:
      type: object
      required: [id, unitUsahaId, nama, kategori, hargaBeli, hargaJual, stok, satuan, createdAt]
      properties:
        id:
          type: integer
        unitUsahaId:
          type: integer
        nama:
          type: string
        kategori:
          type: string
        hargaBeli:
          type: number
        hargaJual:
          type: number
        stok:
          type: number
        satuan:
          type: string
        createdAt:
          type: string
    ProdukInput:
      type: object
      required: [unitUsahaId, nama, kategori, hargaBeli, hargaJual, stok, satuan]
      properties:
        unitUsahaId:
          type: integer
        nama:
          type: string
        kategori:
          type: string
        hargaBeli:
          type: number
        hargaJual:
          type: number
        stok:
          type: number
        satuan:
          type: string
    ProdukUpdate:
      type: object
      properties:
        nama:
          type: string
        hargaBeli:
          type: number
        hargaJual:
          type: number
        stok:
          type: number

    # ─── TRANSAKSI ────────────────────────────────────────────────────────────
...
    LaporanKeuangan:
      type: object
      required: [koperasiId, tahun, bulan, totalPemasukan, totalPengeluaran, labaRugi, totalSimpanan, totalPinjaman, totalAset]
      properties:
        koperasiId:
          type: integer
        tahun:
          type: integer
        bulan:
          type: ["integer", "null"]
        totalPemasukan:
          type: number
        totalPengeluaran:
          type: number
        labaRugi:
          type: number
        totalSimpanan:
          type: number
        totalPinjaman:
          type: number
        totalAset:
          type: number
        rincianUnit:
          type: array
          items:
            $ref: "#/components/schemas/RincianUnit"
    RincianUnit:
      type: object
      required: [unitUsahaId, namaUnit, omzet, labaUnit]
      properties:
        unitUsahaId:
          type: integer
        namaUnit:
          type: string
        omzet:
          type: number
        labaUnit:
          type: number
    LaporanShu:
      type: object
      required: [koperasiId, tahun, totalShu, shuPerAnggota]
      properties:
        koperasiId:
          type: integer
        tahun:
          type: integer
        totalShu:
          type: number
        shuPerAnggota:
          type: array
          items:
            $ref: "#/components/schemas/ShuAnggota"
    ShuAnggota:
      type: object
      required: [anggotaId, namaAnggota, totalSimpanan, totalBelanja, bagianShu]
      properties:
        anggotaId:
          type: integer
        namaAnggota:
          type: string
        totalSimpanan:
          type: number
        totalBelanja:
          type: number
        bagianShu:
          type: number
    LaporanAudit:
      type: object
      required: [koperasiId, tahun, temuan, rekomendasi, skorKesehatan]
      properties:
        koperasiId:
          type: integer
        tahun:
          type: integer
        temuan:
          type: array
          items:
            type: string
        rekomendasi:
          type: array
          items:
            type: string
        skorKesehatan:
          type: number
        statusKeuangan:
          type: string

    # ─── DASHBOARD ────────────────────────────────────────────────────────────
...