# lib/api-zod/src/generated/types/rincianUnit.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Sistem Koperasi Merah Putih API
 * OpenAPI spec version: 0.1.0
 */

export interface RincianUnit {
  unitUsahaId: number;
  namaUnit: string;
  omzet: number;
  labaUnit: number;
}